% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{asMatrix}
\alias{asMatrix}
\alias{asMatrix.default}
\alias{asMatrix.matrix}
\alias{asMatrix.data.frame}
\title{Convert to dense matrix}
\usage{
asMatrix(object, ...)

\method{asMatrix}{default}(object, chunk.size = 1000, verbose = TRUE)

\method{asMatrix}{matrix}(object, ...)

\method{asMatrix}{data.frame}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods.}

\item{chunk.size}{An integer indicating number of rows to convert in each
chunk.}

\item{verbose}{Logical, whether to print the progress information to console.}
}
\value{
A dense \code{matrix} converted from \code{object}.
}
\description{
Convert to dense matrix
}
