% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkPackages}
\alias{checkPackages}
\title{Check if specific packages are installed}
\usage{
checkPackages(packages = character(), required = TRUE, ...)
}
\arguments{
\item{packages}{A character vector including package names to check.}

\item{required}{Logical, whether or not the \code{packages} are necessary.}

\item{...}{Arguments passed to other methods. See \code{\link[=requireNamespace]{requireNamespace()}}.}
}
\value{
When any of \code{packages} is not installed, a warning or an error will be raised
according to \code{required}.
}
\description{
Check if specific packages are installed
}
