% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filterData}
\alias{filterData}
\alias{filterData.logical}
\alias{filterData.character}
\alias{filterData.factor}
\alias{filterData.numeric}
\alias{filterData.data.frame}
\title{Filter data according to set or ranges}
\usage{
filterData(x, ...)

\method{filterData}{logical}(x, ...)

\method{filterData}{character}(x, include = NULL, invert = FALSE, ...)

\method{filterData}{factor}(x, include = NULL, invert = FALSE, ...)

\method{filterData}{numeric}(x, include = NULL, invert = FALSE, ...)

\method{filterData}{data.frame}(x, column, include = NULL, invert = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Arguments passed to other methods.}

\item{include}{One of the following:
\itemize{
\item A character vector indicating the element set to include.
\item A 2-element numeric vector indicating the range to include.
\item A list of several 2-element numeric vectors indicating multiple range
to match. Any vector that doesn't have 2 elements will be ignored.
}}

\item{invert}{Logical. If \code{TRUE} return the logical indices for elements that
are not within \code{include}.}

\item{column}{A string indicating the column names used for filtering.}

\item{verbose}{Logical, whether to print the progress information to console.}
}
\value{
When \code{x} is a logical vector, return itself.

When \code{x} is a character or numeric vector, return a logical indices of the
elements in \code{x} that match \code{include}.

When \code{x} is a \code{data.frame}, return a logical indices of the rows at which the
\code{column} matches \code{include}.
}
\description{
Filter data according to set or ranges
}
