% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{mkdir}
\alias{mkdir}
\title{Create a directory}
\usage{
mkdir(path, chdir = FALSE)
}
\arguments{
\item{path}{A character vector containing a single path name.}

\item{chdir}{Logical, whether to change the work directory into \code{path}.}
}
\value{
An invisibly logical vector indicating if the operation succeeded.
}
\description{
Create a directory with an option to set it as the work directory.
}
