% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{pasteFactors}
\alias{pasteFactors}
\title{Collapse two factors}
\usage{
pasteFactors(x, y, collapse = "_", rev = FALSE)
}
\arguments{
\item{x, y}{Factors to collapse}

\item{collapse}{A character string to separate the results.}

\item{rev}{Logical, whether or not to use \code{y} as the major levels.}
}
\value{
A new factor vetcor that collapses \code{x} and \code{y}.
}
\description{
Collapse two factors
}
