% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{readRDX}
\alias{readRDX}
\alias{Load}
\title{Reload saved datasets}
\usage{
readRDX(file, verbose = TRUE)

Load(file, ...)
}
\arguments{
\item{file}{A character string giving the name of file to load. Must be an
\code{.Rda} or \code{.Rds}.}

\item{verbose}{Logical, whether to print the progress information to console.}
}
\description{
Reload saved datasets
}
