% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{recodeFactor}
\alias{recodeFactor}
\alias{recodeFactor.default}
\alias{recodeFactor.data.frame}
\title{Change factor levels}
\usage{
recodeFactor(x, ...)

\method{recodeFactor}{default}(x, map, keep.order = FALSE, ...)

\method{recodeFactor}{data.frame}(x, map, column, keep.order = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{x}{An object to modify}

\item{...}{Arguments passed to other methods.}

\item{map}{A list mapping new levels to old levels. Must formatted as
following:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  new.level1 = c(old.level1, old.level3, ...),
  new.level2 = old.level2,
  new.level3 = old.level4,
  ...
)
}\if{html}{\out{</div>}}}

\item{keep.order}{Logical, whether or not to keep the orignial order of old
levels. Default is \code{FALSE}, which means the new order should depend on names
of \code{map}.}

\item{column}{Character or integer scalar, indicating which column to be
recoded.}

\item{verbose}{Logical, whether to print the progress information to console.}
}
\value{
A factor or character vector that is recoded according to \code{map}.
}
\description{
Change factor levels
}
