% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{rowMeanPct}
\alias{rowMeanPct}
\alias{rowMeanPct.CsparseMatrix}
\alias{rowMeanPct.matrix}
\title{Calculate the means and percents of values}
\usage{
rowMeanPct(object, ...)

\method{rowMeanPct}{CsparseMatrix}(object, cell.groups, mean.fxn = rowExpMean, min.exp = 0, ...)

\method{rowMeanPct}{matrix}(object, group.by, mean.fxn = rowExpMean, min.exp = 0, ...)
}
\arguments{
\item{object}{A matrix-like object}

\item{...}{Arguments passed to other methods.}

\item{cell.groups}{One of the following:
\itemize{
\item A factor vector containing group labels for columns of \code{object}. The
length must be equal to \code{ncol(object)}.
\item A list containing the column indices (integer or character if \code{object}
has column names) for each group.
}}

\item{mean.fxn}{A function used to calculate the row means of \code{object}.
Default is \code{\link{rowExpMean}} in non-logged space.}

\item{min.exp}{A numeric scalar indicating the minimum value to calculate the
positive percents. Default is 0.}
}
\value{
A list containing three matrices where each row stands for a feature and each
column stands for a group:
\itemize{
\item \code{n.cells}: The numbers of positive observations for each feature within
corresponding groups.
\item \code{avg.exp}: The mean values calculated from \code{mean.fxn}.
\item \code{avg.pct}: The percents of positive observations within corresponding
groups.
}
}
\description{
Calculate the means (in log-spcae) and positive (typically non-zero) percents
for each row in a matrix, where columns are grouping by a factor vector.
}
